+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - (Maintenance) Reduce binary size
 - (Maintenance) Update Features Description, Supported K8S Versions and K8S Dependency to 1.28.5
 - (Feature) (ML) Featurization Job Type
+- (Bugfix) Don't abort plan in case of optional action timeout
 
 ## [1.2.36](https://github.com/arangodb/kube-arangodb/tree/1.2.36) (2024-01-08)
 - (Documentation) Improvements and fixes for rendered documentation (GH pages)